% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sumlog.R
\name{fishers_method}
\alias{fishers_method}
\title{Fishers method for reducing p-values
This function reduces p-values across a data table, per a grouping variable provided as a single column}
\usage{
fishers_method(data, p_value, group)
}
\arguments{
\item{data}{Data frame of p-values, including one column for grouping variable, one column of p-values}

\item{p_value}{Column name for p-values. Accepts string as column name}

\item{group}{Column name for grouping variable. Accepts string as column name.}
}
\value{
Named numeric vector. Names are grouping variable, value is p-value reduced by Fisher's method.
}
\description{
Fishers method for reducing p-values
This function reduces p-values across a data table, per a grouping variable provided as a single column
}
